// backend/src/types/AreaTransactions.ts
export interface AreaTransactions {
  area_id: number | null;
  area_name: string | null;
  villa_tx_all: number | null;
  villa_value_all: number | null;
  villa_tx_last_year: number | null;
  villa_value_last_year: number | null;
  apt_tx_all: number | null;
  apt_value_all: number | null;
  apt_tx_last_year: number | null;
  apt_value_last_year: number | null;
}